Mere scenex.util into scenex.utils
More precisely, it was split into two new private modules.
