+++ src/scenex/model/_nodes/node.py
@@ -382,6 +382,6 @@ class Node(EventedBase):
 
     def tree_repr(self) -> str:
         """Return an ASCII/Unicode tree representation of self and its descendants."""
-        from scenex.util import tree_repr
+        from scenex.utils._tree import tree_repr
 
         return tree_repr(self, node_repr=object.__repr__)
