+++ tests/test_basic_scene.py
@@ -70,7 +70,9 @@ def test_basic_view(basic_view: snx.View) -> None:
     assert isinstance(repr(basic_view), str)
     assert isinstance(basic_view.model_dump(), dict)
     assert isinstance(basic_view.model_dump_json(), str)
-    assert snx.util.tree_repr(basic_view.scene, node_repr=_obj_name) == EXPECT_REPR
+    assert (
+        snx.utils._tree.tree_repr(basic_view.scene, node_repr=_obj_name) == EXPECT_REPR
+    )
     ary = basic_view.render()
     assert isinstance(ary, np.ndarray)
 
@@ -88,9 +90,9 @@ def test_view_tree_matches_native(basic_view: snx.View) -> None:
     structure of the tree generated by the native backend."""
     basic_view._get_adaptors(create=True)
 
-    model_tree = snx.util.tree_dict(basic_view.scene, obj_name=_obj_name)
+    model_tree = snx.utils._tree.tree_dict(basic_view.scene, obj_name=_obj_name)
     native_scene = _native_scene(basic_view.scene)
-    view_tree = snx.util.tree_dict(native_scene, obj_name=_obj_name)
+    view_tree = snx.utils._tree.tree_dict(native_scene, obj_name=_obj_name)
     assert isinstance(view_tree, dict)
     assert model_tree == _filter_backend_nodes(view_tree)
